Move trackers/tracker.py into utilities/tracker.py
