+++ wikimedia/executors/downloader.py
@@ -8,10 +8,9 @@ import tempfile
 
 import magic
 import requests
-from trackers.tracker import Result, Tracker
 from utilities.exceptions import DownloadException
-
-from wikimedia.utilities.helpers import S3Helper
+from utilities.helpers import S3Helper
+from utilities.tracker import Result, Tracker
 
 
 class Downloader:
